+++ internal/command/fetchChannelMembersInfo/handler.go
@@ -5,10 +5,12 @@ import (
 	"fmt"
 	"github.com/revilon1991/tg-parser/internal/config"
 	"github.com/revilon1991/tg-parser/internal/useCase/getMembers"
+	"github.com/revilon1991/tg-parser/internal/useCase/getUser"
 	"github.com/urfave/cli"
 	"io/ioutil"
 	"log"
 	"net/http"
+	"time"
 )
 
 func Handle(c *cli.Context) {
@@ -41,10 +43,60 @@ func Handle(c *cli.Context) {
 
 		_ = json.Unmarshal(responseMembers, &responseMemberList)
 
-		saveMembers(responseMemberList)
+		memberInfoList := MemberInfoList{
+			MemberList:        make(map[int]Member, 0),
+			ChannelExternalId: channel.ChannelId,
+			ChannelId:         channel.Id,
+		}
 
-		memberList := getMemberIdList(responseMemberList)
+		for userExternalId, responseMember := range responseMemberList.Members {
+			joinChannelDate := time.Unix(int64(responseMember.JoinedChatDate), 0).Format("2006-01-02 15:04:05")
 
-		saveRelationChannelMember(channel, memberList)
+			user := getUserInfo(responseMember.UserId)
+
+			memberInfoList.MemberList[userExternalId] = Member{
+				ChannelId:       channel.Id,
+				JoinChannelDate: joinChannelDate,
+				UserExternalId:  responseMember.UserId,
+				Username:        user.Username,
+				FirstName:       user.FirstName,
+				LastName:        user.LastName,
+				PhoneNumber:     user.PhoneNumber,
+				Type:            user.Type,
+				Bio:             user.Bio,
+			}
+		}
+
+		saveMembers(memberInfoList)
+		fetchMemberIdList(memberInfoList)
+		saveRelationChannelMember(memberInfoList)
 	}
+}
+
+func getUserInfo(userId int) getUser.User {
+	url := fmt.Sprintf(
+		"http://localhost:8080%s?user_id=%d",
+		config.Routing.V1GetUser,
+		userId,
+	)
+
+	res, err := http.Get(url)
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	responseMember, err := ioutil.ReadAll(res.Body)
+
+	err = res.Body.Close()
+
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	user := getUser.User{}
+
+	_ = json.Unmarshal(responseMember, &user)
+
+	return user
 }